requirements.txt
uwsgi
Django
django-apscheduler
django-crispy-forms
django-extra-views
django-filter
django-tables2
APScheduler
pendulum
requests
